Fix invalid photometric tag for grayscale tifs

+++ bindings/csharp/ArtomatixImageLoader/ArtomatixImageLoaderTests/TestAImg.cs
@@ -120,7 +120,7 @@ namespace ArtomatixImageLoaderTests
                         img2.decodeImage(data2);
 
                         for (int i = 0; i < data.Length; i++)
-                            Assert.AreEqual(data [i], data2 [i]);
+                            Assert.AreEqual(data[i], data2[i]);
                     }
                 }
             }
@@ -210,8 +210,7 @@ namespace ArtomatixImageLoaderTests
             });
         }
 
-
-        public static void TestWriteIMG<T>(int width, int height, AImgFormat format, AImgFileFormat fileformat, float allowedDelta = 0) where T : struct
+        public static void TestWriteIMG<T>(int width, int height, AImgFormat format, AImgFileFormat fileformat, float allowedDelta = 0, bool deleteAfterwards = true) where T : struct
         {
             var img = new AImg(fileformat);
             T[] data = new T[width * height * format.numChannels()];
@@ -221,20 +220,15 @@ namespace ArtomatixImageLoaderTests
             for (int i = 0; i < data.Length; i++)
             {
                 if (format.bytesPerChannel() == 2)
-                    data [i] = (T)Convert.ChangeType(r.Next(0, ushort.MaxValue), typeof(T));
+                    data[i] = (T)Convert.ChangeType(r.Next(0, ushort.MaxValue), typeof(T));
                 else if (format.bytesPerChannel() > 1)
                     data[i] = (T)Convert.ChangeType(r.Next(0, 255) / 255.0f, typeof(T));
                 else
                     data[i] = (T)Convert.ChangeType(r.Next(0, 255), typeof(T));
-
-
             }
 
             using (var f = new FileStream(getImagesDir() + "/testOut", FileMode.Create))
                 img.writeImage<T>(data, width, height, format, f);
-
-
-
 
             T[] readBackData = null;
             using (AImg f = new AImg(new FileStream(getImagesDir() + "/testOut", FileMode.Open)))
@@ -246,14 +240,15 @@ namespace ArtomatixImageLoaderTests
             for (int i = 0; i < data.Length; i++)
                 Assert.That(data[i], Is.EqualTo(readBackData[i]).Within(allowedDelta));
 
-
-            try
+            if (deleteAfterwards)
             {
-                Directory.Delete(getImagesDir() + "/testOut");
+                try
+                {
+                    Directory.Delete(getImagesDir() + "/testOut");
+                }
+                catch { }
             }
-            catch { }
         }
-
 
         [Test]
         public static void TestWriteTiffs()
@@ -262,6 +257,7 @@ namespace ArtomatixImageLoaderTests
             TestWriteIMG<ushort>(128, 128, AImgFormat.RGBA16F, AImgFileFormat.TIFF_IMAGE_FORMAT);
             TestWriteIMG<byte>(128, 128, AImgFormat.RGBA8U, AImgFileFormat.TIFF_IMAGE_FORMAT);
             TestWriteIMG<ushort>(128, 128, AImgFormat.RGBA16U, AImgFileFormat.TIFF_IMAGE_FORMAT);
+            TestWriteIMG<ushort>(128, 128, AImgFormat.R16U, AImgFileFormat.TIFF_IMAGE_FORMAT);
         }
 
         [Test]
@@ -288,7 +284,6 @@ namespace ArtomatixImageLoaderTests
             TestWriteIMG<byte>(128, 128, AImgFormat.RG8U, AImgFileFormat.TGA_IMAGE_FORMAT);
             TestWriteIMG<byte>(128, 128, AImgFormat.R8U, AImgFileFormat.TGA_IMAGE_FORMAT);
         }
-
 
         [Test]
         public static void TestWrite2ChannelPNGs()
